server/requirements.txt
brotli==1.0.9
email-validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
gunicorn==23.0.0
httptools==0.6.4
importlib-metadata==8.0.0
inflect==7.3.1
itsdangerous==2.2.0
jaraco-functools==4.2.1
jaraco.collections==5.1.0
jinja2==3.1.6
langchainhub==0.1.16
miniaudio==1.59
openai-agents==0.2.7
opencv-python==4.9.0.80
orjson==3.11.1
pip-chill==1.0.3
platformdirs==4.2.2
pyaudio==0.2.14
pydantic-extra-types==2.10.5
pygame==2.5.2
pysocks==1.7.1
python-multipart==0.0.9
retell-sdk==4.40.0
pyyaml==6.0.2
ruamel.yaml==0.17.21
tomli==2.0.1
twilio==9.2.1
ujson==5.10.0
uuid==1.30
watchfiles==1.1.0
websockets==15.0.1